12. SEO_Content_Auto_QA_Tool/requirements.txt
playwright==1.42.0
beautifulsoup4>=4.12.2,<4.13.0
pandas>=2.1.4,<2.3.0
sqlalchemy==2.0.28
click==8.1.7
rich>=13.7.0,<13.8.0
pytest==8.0.2
pyyaml==6.0.1
requests==2.31.0
python-dotenv>=1.0.0,<2.0.2
aiohttp==3.9.3
lxml==5.1.0 